Add `raw_request` method to Client and AsyncClient
Sends a raw request with the underlying httpx Client.

This method ignores a set api_url and allows send request to custom
hosts. The raw httpx response will be returned.
If `apply_auth_flow` keyword arg is set to `True`, the Authenticator
auth_flow (sign request) will be applied to the request.

If `apply_cookies` keyword arg is set to `True`, website cookies from
Authenticator will be applied to the request.