update get_next_action_from_soup()
When using a custom captcha callback, get_next_action_from_soup() was not returning the right `form` from the `login_soup`, and then would crash due to a `KeyError` when it wouldn't return the first `form` found in the soup, that didn't have a key of either `'method'` or `'action'`. 

The `or` operator had some strange side effects where it was returning the second `form` found in a return for the captcha. Debugging it using "soup.find_all()" on the `login_soup` would return multiple form items, and for some reason even when `search_field` was set to `None`, it would return the form found at index 1 instead of index 0.