fully support pre-Amazon accounts
The `Authenticator` now holds the `with_username` attribute and other
stuff for pre-Amazon accounts.

Auth data saved with v0.8 or later can not be loaded with versions
less than v0.8!
If an auth file for an pre-Amazon account (with_username=True) was
created with v0.7.1 or v0.7.2 set `auth.with_username` to `True` and
save the data again.
After this, deregistration, refreshing access tokens and requesting
cookies for another domain will work for pre-Amazon accounts.