V0.8.0dev (#89)
* fully support pre-Amazon accounts

The `Authenticator` now holds the `with_username` attribute and other
stuff for pre-Amazon accounts.

Auth data saved with v0.8 or later can not be loaded with versions
less than v0.8!
If an auth file for an pre-Amazon account (with_username=True) was
created with v0.7.1 or v0.7.2 set `auth.with_username` to `True` and
save the data again.
After this, deregistration, refreshing access tokens and requesting
cookies for another domain will work for pre-Amazon accounts.

* setdefault `with_username` to `False` when instantiate an Authenticator

* rework setdefault `with_username`

* Update client.py

- `Client` and `AsynClient` now accepts session kwargs which are bypassed to the underlying httpx Client
- rename (and rework) `Client._split_kwargs` to `Client._prepare_params`

* Update CHANGELOG.md

* Add `response_callback` to `Client` and `AsyncClient`

This allows to set a custom response callback like:

```python
def custom_response_callback(resp):
    return resp
```

This returns the unprepared `httpx.Response` object for further purposes (e.g. accessing respopnse headers).

* remove trailing newlines

* Update activation_bytes.py

* remove whitspaces

* work on docs and docstring

* Update CHANGELOG.md

* Create example to get and analyse listening statistics

Include an example that authenticates using the browser and goes on to extract aggregate per month listening time since account creation (currently hardcoded to 2013) to the current month. 

Then the script uses time-series analysis to find trends and seasonalities in the listening pattern.

* Create download_bookmarks.py

Create an example to download all bookmarks for all books in the library and store them in a json file 'books.json'.

* rework `_prepare_api_path` method of client classes

- absolute url paths are passed through without preparation
- query params for a given url will be keep intact

* Update CHANGELOG.md

Co-authored-by: Botty Dimanov <bottydim@users.noreply.github.com>